Webclient: fix view config "Rows" default (#208).
This (client-side) error does not accur anymore:
{{{
    TextRendererError: field "xx" has no "rows" setting
}}}

If no "Rows" value is set (or if it is 0) then 1 is used as the default.

The "Rows" setting is renamed:
    "Rows" -> "Input Field Rows"

BREAKING CHANGES

1 view config URI is changed:
    "dm4.webclient.rows" -> "dm4.webclient.input_field_rows"

In your plugins you must perform a search/replace accordingly.
Especially your JSON migration files might be affected, e.g.
{{{
    view_config_topics: [
        {
            type_uri: "dm4.webclient.view_config",
            composite: {
                dm4.webclient.rows: 15      # -> dm4.webclient.input_field_rows: 15
            }
        }
    ]
}}}

NOTE 1: your custom renderers might need to be adapted. The "rows" property of the "page_model" object that is passed to the renderer functions is renamed to "input_field_rows".

NOTE 2: an **association definition** can no longer override a child type's view configuration setting with a **falsish** value (0, "", false). This is regarded as not needed anymore because the semantics of the former "Viewable" and "Editable" settings is inverted (see last commit).

See ticket 208.