Core fix: no doublettes in comp val model (#339).
BREAKING CHANGE

CompositeValueModel public API:
    semantics of add() method has slightly changed:
    If the topic is already added (based on ID-equality) nothing is performed.

This change is required in conjunction with plugins which enrich a topic model in response to a PRE_SEND event, e.g. by fetching multi-facets. For an update request the facets might be fetched already and would be added again through PRE_SEND yielding to doublettes in the model (and eventually to double renderings in the webclient). This happened e.g. with the Kiezatlas plugin.

This Kiezatlas error does not occur anymore: multi-facet values are rendered twice after form submission.

Your plugins are probably *not* required to be updated.

See ticket 339.