Webclient: fix tooltip text (#421).
The tooltips of topics in lists (detail panel) now read e.g. "Person" instead of "A Person".
The "A" does not work for type names in languages other than English.

Furthermore the tooltip shows the topic's URI (if there is one).
This is helpful to distinguish e.g. types when they have the same name (e.g. there are 3 types named "Description").

Close ticket 421.