New Core Event PRE_PROCESS_REQUEST (#480, #478).
Fired before a request is processed. That is *before* a possible resource method is invoked.
This allows a plugin e.g. to abort request processing in case a pre-condition is not met (as stated e.g. by the `If-Modified-Since` or `If-Unmodified-Since` headers).

New listener interface:
    de.deepamehta.core.service.event.PreProcessRequestListener
Listener method:
    void preProcessRequest(ContainerRequest request)

Note: ContainerRequest is a Jersey-specific class
(com.sun.jersey.spi.container.ContainerRequest).
We should remove the Jersey dependency and move to JAX-RS 2.0 instead.

Close ticket 480.
See ticket 478.