Time plugin: enrich with timestamp (#386, #479).
If the response contains a single DeepaMehtaObject its composite model is enriched with a `dm4.time.modified` entry. The value is a UNIX timestamp. If the modification date is unknown the value is -1.

This way an object's timestamp is accessible at the Webclient application layer (client-side). It could be used for display or for setting the `If-Unmodified-Since` header in a conditional PUT request.

In preparation of #479 (Edit Conflict Detection).

See ticket 386.
See ticket 479.