Caching plugin: edit conflict detection (#479).
The server rejects a PUT request when the resource has changed meanwhile. In this case the response is 412 (Precondition Failed). The check is based on the resource's "last modified" timestamp.

The Webclient makes a PUT request conditional by adding a `If-Unmodified-Since` header to it.

So the "edit conflict detection" feature is basically available. Issues remain.

2 requests are affected for the moment:
{{{
PUT /core/topic/{id}
PUT /core/association/{id}
}}}

See ticket 479.