Core: drop Felix ExtHttpService (#448).
The Access Control plugin doesn't register a servlet filter anymore but listens to the PRE_PROCESS_REQUEST DeepaMehta core event.

BREAKING CHANGE

PluginActivator's registerFilter() method is dropped.

IMPORTANT

This solution works for a standard DM installation but is not fully compatible with DM's previous request filtering behavoir. In particular a closed DM installation (read_requires_login=true) have a serious issue.

It turned out that request filtering via a servlet filter and request filtering via the PRE_PROCESS_REQUEST core event are **not** equivalent. While a servlet filter sees **all** requests a PreProcessRequestListener sees only the requests bound to the **Jersey servlet** but not the requests to **static resources**.

As a consequence a closed DM installation is not fully closed anymore, and in case of Safari is not working at all (Safari brings up no login dialog for a 401 response when the request was sent via XHR).

So, this commit must not meet the DM production line. Further work is required. We could either 1) further adapt the DM core in the way the PRE_PROCESS_REQUEST event is fired for static resource requests as well, or we could 2) investigate weather Pax Web provides a working custom API for deploying servlet filters (an equivalent to Felix's ExtHttpService).

See #448.