Pax Web success! (#448).
The REST services (the Jersey servlet) is working.
Finally DM runs with Pax Web!

The solution was to register the Jersey servlet at "/*" instead of "/" (which I think is not OSGi compliant and does not work with Felix HTTP Jetty as well). So a single asterisk character did the trick.

You can test it in the "pax-web" branch.

What is missing now is the adaption of the SSL configuration parameters.

BTW: with Pax Web we finally have the opportunity to set the Jetty webserver to listen on a specific network interface (see #39).

You might not consider this possible: yesterday I bought a bottle of spaish red wine called Asterisk. The label is all blank but with a single * on it.

See #448.