Webclient: HTML based topic rendering pt.1 (#505).
Begin of a new concept: Topics are rendered as HTML, instead of being drawn on the canvas. Thus drawing, event handling, object click detection relies on the native browser capabilities. This simplifies the development of DeepaMehta View Customizers tremendously.

This is in experimental stage. There is a new branch "topic-html". Not fully usable.

Works in conjunction with the Box Renderer "topic-html" branch.

See #505.