Webclient: sort listing by topic type (#315).
The "what's related" topic listing in the detail panel is grouped by Topic Type (instead Association Type).
Thus the group headings are more user friendly.

BREAKING CHANGE

Server-side serialization format: a RelatedTopic's association is NOT serialized anymore. That is the JSON object has no "assoc" property anymore.
For the moment it is not clear weather the association will be needed in the future. For the moment it reduces network traffic considerably.

See #315.