Expand canvas view facade by update_topic (#505).
View Customization Framework:
The canvas view facade (as passed to a view model customizer) has a new method
{{{
update_topic(topic)
}}}
A view customizer calls this method to update the view e.g. once topic childs are loaded in response of clicking an expansion handle.

It updates the TopicViewmodel and the TopicView according to the changed domain topic, fires the "on_update_topic" customizer hook, renders the canvas, and (in case of a DOM based customizer) repositions the topic DOM.

See #505.