Webclient: "Create" menu on canvas (#529).
Right-clicking the canvas brings up a Create menu with the same topic types as listed in the toolbar's Create menu. The created topic is positioned where the right-click is performed.

The "+1" button near the toolbar's Create menu is gone.

DEVELOPER NOTES

Client-Side Plugin Framework:

Plugins can place disabled items in the topic/association/canvas context menus. They do so in the respective `.._commands` hook by setting a command's `disabled` property to true.

Plugins can add icons to the items in the topic/association/canvas context menus. They do so in the respective `.._commands` hook by setting a command's `icon` property to the (absolute or relative) URL of the icon file.

GUI Toolkit:

Menu items added by the `add_item()` method can have a boolean `disabled` property. If true the menu item appears as disabled in the menu. This works for Menu, ContextMenu, and Combobox.

See #529.