Reveal direct childs via page panel (#315).
The direct child topics of the selected topic (which are not appearing anymore in the page panel's related topics listing) can be revealed via the page panel's upper area (which shows the nested boxes): when hovered the corresponding box it is highlighted and a pointer cursor appears. On click the underlying topic is revealed in the map.

See #315.