Caching module: fix retyping (#547).
CHANGES

The Caching module's strategy for determining the last modification timestamp known at client-side (to be used for the `If-Unmodified-Since` request header) when performing a PUT request is extended:

1. if the object contained in the update request has already a timestamp (in its `composite` property) that one is used.

2. if not, it is checked weather the object contained in the update request is the same as currently displayed in the page panel (based on ID match), the page panel object's timestamp is used.

3. if the IDs do not match an exception is thrown at client-side and the request is not send.

New is step 2.
This means the caller of dm4c.update_topic() or dm4c.do_update_association() is not required to include a timestamp in the object to be updated as long as this object is currently displayed in the page panel.

`dm4c.page_panel.get_displayed_object()` is again a public method.

See #547.