Webclient: add "Create Note" button (#529).
In the toolbar, beside the Create menu, there is a "Create Note instantly" button. It can be triggered via Alt-n. The button creates a Note topic instantly. It is useful for "in a hurry" situation when one want enter information right away.
Thank you, Malte, for bringing up this use case!

See #529.