Editor for aggregated composites (#337).
You can now have aggregated composites in your data model and can edit such complex topics in the Webclient's detail panel. However, the current implementation differs from what is described in the original ticket. The current implementation is pragmatic, the usage is intuitive, and no proliferating topic copies are created. For example you can edit a Person's address right in the Person form, even if Address would be defined via Aggregation Definition.

"semantic side effects" (see comments in #30) are possible in principle but only when associations are manipulated manually. In the scenario of an aggregated Address you can assign the same Address instance to 2 persons by manually creating an association. (Note: when editing via detail panel one Address instance is currently never assigned to 2 Persons.) In that case, when you edit the Address via Person 1 in the page panel, it changes for Person 2 as well. You can consider that as a "semantic side effect".

In some cases that is exactly what you want, in others it is not. However, that's the way the Webclient works for now. Regarding development effort it reflects the most pragmatic solution. The underlying semantic problem is conceptually huge, and in the way I see it there no general solution exists except further complicating the meta model and/or the GUI, which I feel is not desirable.

Future strategy:
- The GUI could be extended in the way described in the ticket. It could support the creation of aggregated composites that are actually shared between various parents.
- Different form rendering and processing behavior is realizable by custom page renderers.

See #337.