Core: sync view on Include in Label change (#564).
When changing the "Include in Label" setting in a type form while the respective association definition is visible on the canvas, the association label on the canvas is updated.

CHANGES

When calling a `CompositeValue` manipulator method (`set()`, `setRef()`, `remove()`) the passed set of Directives is extended by a UPDATE_TOPIC resp. UPDATE_ASSOCIATION directive.

BREAKING CHANGES

Core API Type interface
    - `setLabelConfig()` has a `directives` parameter

Close #564
See #576