Page Panel: icon vs link click behavior (#579).
In the detail panel's "related topic" lists: when the topic is contained in the current topicmap its icon is surrounded by a thin red line (instead of dimming it). When such an icon is clicked the respective association is revealed (in case it was hidden) and the topicmap is translated to bring the topic into the viewport (in case it was outside the viewport).

CHANGES

`dm4c.render.icon_link()` got a 4th parameter: `css_class` - the class(es) to add to the <img> element.

See #579.