Caching fix: regard 0 as valid timestamp (#547).
The Caching module regards 0 as a valid timestamp. Note: the server sends 0 as the default timestamp for objects which don't have a timestamp. This happens e.g. when upgrading from a DM 4.1 installation. Timestamps were introduced only in DM 4.1.1

Thank you very much, Malte, for testing the upgrade scenario!

See #547