GUIToolkit fix: clear DOM on close dialog (#583).
Once a dialog is closed it is removed from DOM, regardless what close gesture is used.
The DOM is not polluted.

BREAKING CHANGES

Only "one-shot" dialogs are still suppurted. That means a plugin must add the dialog to the DOM only when it is supposed to appear on-screen.
In contrast, "static" dialogs are not supported aynmore (that was when a plugin adds the dialog to the DOM on init time and reuses it many times).

GUIToolkit Dialog API:

- `open()`    is dropped. The dialog is now opened automatically on creation time.
- `close()`   closes the dialog and removes it from the DOM.
- `destroy()` is dropped. Use close() instead.
- `empty()`   is dropped.
- `append()`  is dropped.

Webclient API:

- `dm4c.do_delete_topic()`       takes a **topic ID** as argument (instead a topic).
- `dm4c.do_delete_association()` takes a **topic ID** as argument (instead an association).
- `dm4c.do_retype_topic()`       takes a **topic ID** as argument (instead a topic).

See #583.