Access Control: assoc READ permission (#592).
For the first time a public topicmap which contains partly private objects is properly displayed.

When the user is not logged in the private parts are not displayed, and neither appear in the detail panel.
Once logged in, the private parts appear.

Note: an association is granted READ permission only if the user has READ permission to all the 3 objects: the association itself AND the 2 player topics.

See #592.