Access Ctrl: synthesize CREATE permission (#592).
At client-side the CREATE permission is synthesized from both:
    - is the respective type READable?
    - is the current workspace WRITEable?

Note: at client-side the CREATE permission is one factor to decide which types to be shown in the Create menus.

BREAKING CHANGE

At server-side there is no notion of a CREATE permission anymore.
The `Operation.CREATE` enum constant is removed.

See #592.