Webservice: serialize number return values (#682).
When a plugin's service method returns a number it serializes automatically. The number's string representation is written to the message body, followed by the line separator string.

The plugin developer is no longer required to implement a wrapper class for the return value or a JAX-RS provider class (a message body writer).

See #682.