Geomaps: fix topic enrichment (#684).
Topics going over the wire do not include all their child topics if not requested.

This former bug -- which was caused by the dm4-geomaps module and existed since DM 4.1 -- resulted in crucial network overhead, that is more data than retuested was sent on a regular basis.

IMPORTANT: make sure your plugin does not rely on that buggy behavior, and possibly adapt it.

See #684.