Core: easify static resource access (#692).
`PluginActivator` has a new method:
{{{
InputStream getStaticResource(String name)
}}}
Use this to access a static resource of your plugin.
The argument is the resource name: a "/"-separated path name, relative to the plugin bundle's root directory. It may or may not begin with "/" (it makes no difference).

So, you don't need to call `dms.getPlugin(yourPluginUri)` anymore.
And you don't need to catch any exception.

See #692.