Remove ClientState parameter from Core API (#696).
BREAKING CHANGES

The `clientState` parameter is completely removed from all APIs.
So, just drop it from all your method calls.

If your plugin needs access to the cookies inject `HttpHeaders` and construct a `Cookies` object on it.
`Cookies` is a new class which is call compatible to the former `ClientState` (which is dropped).

Example:
{{{
import de.deepamehta.core.service.Cookies;

import javax.ws.rs.Path;
import javax.ws.rs.core.Context;
import javax.ws.rs.core.HttpHeaders;

@Path("...")
public class YourPlugin ... {

    @Context
    private HttpHeaders httpHeaders;

    void someMethod() {
        Cookies cookies = new Cookies(httpHeaders);
        ...
    }
}
}}}

Note: injecting the HTTP headers only works in a root resource class, that is a class which has a `@Path` annotation.

IMPORTANT: the DM standard distro is not yet working, in particular the Workspaces plugin. This is because we can't cheat anymore by passing `null` as `clientState`. The Workspace plugin did this to avoid the vicious circle that occurs when an workspace assigment (an `Aggregation`) gets itself an workspace assignment, which gets itself ...

To solve this situation we aim to solve #341 ("Association Definitions with custom Association Types") first. We will introduce a dedicated association type `Workspace Assignment` then. This allows the Workspace plugin to skip workspace assignment in case a `Workspace Assignment` association is detected. A generic `Aggregation` in contrast can not be distinguished ... ooh, hey, let's wait a minute ... while I'm writing this the solution becomes evident: of course an Aggregation can be recognized as a workspace assignment: by checking the involved players! OK, lets go this way. Lets wait #341 a little.

See #696.