Core API wording: rename CompValue pt.3 (#702).
BREAKING CHANGE

REST APIs: the `fetch_composite` query parameter is renamed to `include_childs`.

Example:
{{{
GET /core/topic/1234?include_childs=true    ### "include_childs" was "fetch_composite"
}}}

Plugin developers are supposed to carefully update client-side code.

NOTE: the former `fetch_relating_composite` query parameter to include the (relating) association's child topics in the response is not supported at the moment. Do we actually need it in the REST APIs? If so, it should be renamed to `include_assoc_childs` and be implemented in the Core's `JerseyResponseFilter` (analogous to `include_childs`).

See #702.