Core: replace Felix HTTP by Pax Web (#448).
Felix HTTP Jetty is now replaced by Pax Web once and for all (hopefully).

We have Pax Web 3.1.2 with a Jetty 8 (that means Servlet API 3.0).

Also we have Felix 4.4 (that means OSGi 5) now (see #703).

We're still on JDK 6.

DM with Pax Web can be tested on the "pax-web-3" branch.
This is merged into master soon and will be part of the upcoming DM 4.4 release.

See #448.