Fix: permission for creating a workspace (#592).
The "New Workspace..." command appears only if the user has the respective permission.
The rule is simple: every logged in user is allowed to create workspaces.
This rule is different from the rule for regular topics and topicmaps as workspaces are top-of-hierarchy objects.
They are NOT created in the context of another workspace.

See #592.