Add "modifier" info to each topic/assoc (#749).
Every time a topic/association is updated DM attaches the "modifying user" (a username) to the topic/association.
So, DM can tell which user did the most recent modification on a topic/association.

Access Control API

New method:
{{{
String getModifier(DeepaMehtaObject object)
}}}
Returns the modifier of a topic or an association, or null if no modifier is set.

See #749.