AcssCtrl: set modifier user also on create (#749).
The "modifying user" (a username) is also set on topic/association creation time (besides update time).

This is analogous to the Time module which sets the modification timestamp already on creation time. (This in turn is required for the caching mechanism which expects each topic/association having a modification timestamp.)

So, when there is a modification timestamp the corresponding modifying user is supposed to be known as well.

See #749.