Fix memberships (#751).
The user who creates an User Account does *not* become a member of the new user's private workspace.

Underlaying general change:
When a workspace is created the current user does *not* become a member of that workspace anymore. This is not necessary as the current user already becomes the *owner* of that workspace. (A membership is pointless when I'm the owner anyway.)
Note: creating an User Account is a special case in this context. Here, of course, the *new* user becomes the owner of hers private workspace (not the user who creates the account, and thus the private workspace).

See #751.