Access Control Mi#6: convert user accounts (#592).
In case of updating an existing installation to DM 4.5: this migration converts the user accounts. That is the private workspaces are created and the various user account parts are assigned to the proper workspaces.

For the first time search is working on an updated-to-4.5 installation. Issues remain.

Access Control API:

1 new public constant in `AccessControlService`:
{{{
static final String DEFAULT_PRIVATE_WORKSPACE_NAME = "Private Workspace";
}}}

See #592