Access Control: refactor migrations (#592, #324).
In case of updating an existing installation to DM 4.5: this migration reconstructs the user's memberships by the means of 4.5's Membership associations. This works in particular when several workspaces and user accounts have been created in the pre-4.5 installation.

The Access Control migrations are refactored: the plugin makes no use anymore of both, the `postInstall()` hook, and the `allPluginsActive()` listener. These actions performs as migrations now, which makes the control flow and code more clear.

See #592
See #324