Fix workspace assignment check (#592).
The check to which workspace an object is assigned to is based on the `dm4.workspaces.workspace_id` property -- not the Aggregation association. Technical relevant is the property. The association is there just for user benefit (that is visualization).

BREAKING CHANGE

Workspaces API

1 changed method:
{{{
boolean isAssignedToWorkspace(long objectId, long workspaceId)
}}}
It takes an object ID (that is a topic ID or association ID) as the 1st parameter (formerly a `Topic` object).
It can also be used to check the assignment of an association.

See #592.