Update Webclient GUI on retype Membership (#592).
When a user leaves a workspace (by *retyping* a Membership where she is involved in) the Webclient GUI is updated immediately in order to reflect the decreased authority. Pressing the Reload button is not necessary.

Webclient API

2 changed client-side events:
{{{
"post_update_topic"
"post_update_association"
}}}
A 2nd argument is now passed to the listener: the Topic/Association object as it was before the update.

1 new client-side event:
{{{
"pre_update_association"
}}}
1 argument is passed to the listener: the new association model.

See #592.