Webclient GUIToolkit: Submenus. No success (#758).
The GUIToolkit's menus and the underlying Webclient commands model has been extended in order to support submenus resp. subcommands. But without success.

The problem is in jQuery UI in conjunction with the Webclient's "desktop style" UI, in particular the behavior when menus exceed the browser window bottom. As you know the Webclient's overall objective is to avoid the browser window from scrolling. Instead the Webclient trims the menu and scrolls just the menu. Realizing this behavior for the submenus as well is much too hard. Such a behavior should be offered by jQuery UI out-of-the-box. At the moment I see no reasonable solution based on jQuery UI.

The current state is available in the "submenus" branch (which will probably not be continued/merged). The Workspaces plugin adds a "Assign to" command to the topic context menu with a submenu listing the workspaces.

See #758.