Access Control: add "System" authority (#764).
Code that is invoked by the system itself -- as opposed to be invoked by ''a user'' -- runs under the System authority. The System authority has unrestricted access to the database. The System authority is in effect for all code that executes ''outside a request scope'', that comprises the ''migrations'' in particular.

See #764.