Webclient: don't create empty child topics (#171).
When an input field is left empty DM does not create an empty topic for that field. Now this applies not only for combobox fields (child via Aggregation) [01cf855] but also for regular fields (child via Composition).

Furthermore if a regular field is emptied DM deletes the underlying child topic. The objective is to never store empty (simple) child topics in the DB.

Committed to master branch.

CHANGES

Webclient

Number fields can be empty. The Webclient does not fill in 0 automatically anymore. This design decision was made because 0 and "not set" might mean something different to the user.

Core

REST API: in an update request a composition-one child can be removed by using the del_id: prefix. Now the del_id: prefix is supported for all the 4 cases, that is composition/aggregration, one/many.

Core API: analogously `TopicDeletionModel` objects can be present in `ChildTopicModel` objects used for updating composition-one childs. DM deletes the that child topic then.

See #171.