Webclient: use assoc type as field label (#341).
When a custom association type is set and the child type is simple the detail panel renders the name of the custom association type as field label (instead the name of the child type).

Example: child type is "Date" (simple) and custom association type is "Birthday". The date field is labeled "Birthday" then.

Committed to the inst-level-assoc-type branch.

See #341.