REST API: add "include_assoc_childs" param (#659).
The REST API supports 1 new query parameter:
{{{
include_assoc_childs=true|false
}}}
It can be used for every resource method which returns a `DeepaMehtaObject` or an `Iterable` of `DeepaMehtaObject`. If set to `true` the child topics of the object's ''Relating Associations'' are included in the response. This works only if `include_childs=true` is present as well.

BREAKING CHANGES

Core API

1 method removed from `ChildTopics`:
{{{
Iterable<String> childTypeUris()
}}}
Instead a `ChildTopics` object is itself an `Iterable<String>` now (analogous to `ChildTopicsModel`).

See #659.